dddd

+++ App.Core.ApplicationService/ApplicationSerrvices/GenreMovies/GenreMovieService.cs
@@ -16,7 +16,7 @@ namespace App.Core.ApplicationService.ApplicationSerrvices.GenreMovies
         private readonly IMovieRepository<GenreMovie> genreMovieRepository;
         private readonly IMapper mapper;
 
-        public GenreMovieService(IMovieRepository<GenreMovie> _genreMovieRepository,IMapper mapper)
+        public GenreMovieService(IMovieRepository<GenreMovie> _genreMovieRepository, IMapper mapper)
         {
             this.genreMovieRepository = _genreMovieRepository;
             this.mapper = mapper;
@@ -33,44 +33,56 @@ namespace App.Core.ApplicationService.ApplicationSerrvices.GenreMovies
         public async Task<string> Update(GenreMovieUpdateDto item)
         {
             var GMovie = mapper.Map<GenreMovie>(item);
-           await this.genreMovieRepository.Update(GMovie);
-          await genreMovieRepository.Save();
+            await this.genreMovieRepository.Update(GMovie);
+            await genreMovieRepository.Save();
             return $"{item.Id }is update";
             ;
         }
 
         public async Task<int> Delete(int id)
         {
-           var input= genreMovieRepository.GetQuery().Select(x => x.Id != id).FirstOrDefault();
-           if(input==null)
+            var item = genreMovieRepository.GetQuery().FirstOrDefault(x => x.Id == id);
+
+            if (item != null)
+            {
+                await genreMovieRepository.Delete(id);
+                await genreMovieRepository.Save();
+                return id;
+            }
+            else
             {
                 return 0;
             }
-         await genreMovieRepository.Delete(id);
-         await genreMovieRepository.Save();
-            return id;
         }
 
-        public async Task<GenreMovieInputDto> Get(int id)
+        public async Task<GenreMovieOutput> Get(int id)
         {
 
-          var GenresMovies = genreMovieRepository.GetQuery().Select(x => x.Id != id).FirstOrDefault();
-                    
-            var mappedDirectorss = mapper.Map<GenreMovieInputDto>(GenresMovies);
-            List<GenreMovieInputDto> result = new List<GenreMovieInputDto>();
-            result.Add(mappedDirectorss);
-            return mappedDirectorss;
-           
+            var GenresMovies = genreMovieRepository.GetQuery().Select(x => x.Id != id).FirstOrDefault();
+            var mappedGenreMovie = mapper.Map<GenreMovieOutput>(GenresMovies);
+            return mappedGenreMovie;
+
         }
 
-        public Task<List<GenreMovie>> GetAll()
+        public async Task<List<GenreMovieOutput>> GetAll()
         {
-            return genreMovieRepository.GetAll();
+            List<GenreMovie> genreMovies = new List<GenreMovie>();
+
+            genreMovies = genreMovieRepository.GetQuery().ToList();
+            List<GenreMovieOutput> genreMovieOutputs = new List<GenreMovieOutput>();
+
+            foreach (var item in genreMovies)
+            {
+
+                var mappedGenremovies = mapper.Map<GenreMovieOutput>(item);
+                genreMovieOutputs.Add(mappedGenremovies);
+            }
+            return genreMovieOutputs;
         }
 
-        public List<GenreMovie> GetQuery()
-        {
-            return genreMovieRepository.GetQuery().ToList();
-        }
+       // public List<GenreMovieOutput> GetQuery()
+        //{
+          //  return genreMovieRepository.GetQuery().ToList();
+        //}
     }
 }
